+++ src/Command/Doctrine/RollupCommand.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace AvaiBookSports\Bundle\MigrationsMutlipleDatabase\Command\Doctrine;
+
+use Symfony\Component\Console\Input\ArrayInput;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class RollupCommand extends AbstractCommand
+{
+    protected static $defaultName = 'doctrine:migrations:rollup';
+
+    protected function configure() : void
+    {
+        $this
+            ->setAliases(['rollup'])
+            ->setDescription('Rollup migrations by deleting all tracked versions and insert the one version that exists.')
+            ->setHelp(<<<EOT
+The <info>%command.name%</info> command rolls up migrations by deleting all tracked versions and
+inserts the one version that exists that was created with the <info>migrations:dump-schema</info> command.
+
+    <info>%command.full_name%</info>
+
+To dump your schema to a migration version you can use the <info>migrations:dump-schema</info> command.
+EOT
+            )
+            ->addOption('em', null, InputOption::VALUE_REQUIRED, 'Name of the Entity Manager to handle.')
+        ;
+
+        parent::configure();
+    }
+
+    protected function execute(InputInterface $input, OutputInterface $output): int
+    {
+        $newInput = new ArrayInput([]);
+
+        $newInput->setInteractive($input->isInteractive());
+
+        foreach ($this->getDependencyFactories(strval($input->getOption('em'))) as $dependencyFactory) {
+            $otherCommand = new \Doctrine\Migrations\Tools\Console\Command\RollupCommand($dependencyFactory);
+            $otherCommand->run($newInput, $output);
+        }
+
+        return self::SUCCESS;
+    }
+}