+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3]
 ### Added
+- Supported command `doctrine:migrations:execute`
+- Supported command `doctrine:migrations:generate`
+- Supported command `doctrine:migrations:rollup`
 - Supported command `doctrine:migrations:dump-schema`
 - Supported options `--namespace` and `--filter-expression` for `doctrine:migrations:diff` command